requirements.txt
Flask==3.0.3
gunicorn==21.2.0
tensorflow-cpu==2.13.0
numpy==1.26.4
Pillow==10.4.0
Werkzeug==3.0.3
matplotlib==3.8.4
h5py==3.10.0